feat: 프로젝트 참여 기능 구현
- projectId와 일치하는 프로젝트인지 확인
- 참여중인 프로젝트면 toast로 피드백 메시지 출력
- 아니면 participants에 uid : 'Member'로 추가
- user의 projectId에도 projectId를 추가

*추가적으로 dialog들을 파일 두개로 분리하여 관리
