+++ app/lib/src/pages/work_page/work_page.dart
@@ -5,9 +5,9 @@ import 'package:talk_pilot/src/models/project_model.dart';
 import 'package:talk_pilot/src/provider/user_provider.dart';
 import 'package:talk_pilot/src/provider/project_provider.dart';
 
-import 'package:talk_pilot/src/pages/work_page/widgets/work_dialogs.dart';
-import 'package:talk_pilot/src/pages/work_page/widgets/project_card.dart';
 import 'package:talk_pilot/src/pages/project_page/project_detail_page.dart';
+import 'package:talk_pilot/src/pages/work_page/widgets/project_card.dart';
+import 'package:talk_pilot/src/pages/work_page/widgets/project_action_dialog.dart';
 
 class WorkPage extends StatefulWidget {
   const WorkPage({super.key});
